# exec: add cancellation check to operators