--- a/pkg/sql/exec/hashjoiner.go
+++ b/pkg/sql/exec/hashjoiner.go
@@ -386,6 +386,8 @@ type hashTable struct {
 	// differs stores whether the key at any index differs with the build table
 	// key.
 	differs []bool
+
+	cancelChecker CancelChecker
 }
 
 func makeHashTable(
@@ -518,20 +520,21 @@ func (ht *hashTable) finalizeHash(buckets []uint64, nKeys uint64) {
 // computeBuckets computes the hash value of each key and stores the result in
 // buckets.
 func (ht *hashTable) computeBuckets(
-	buckets []uint64, keys []coldata.Vec, nKeys uint64, sel []uint16,
+	ctx context.Context, buckets []uint64, keys []coldata.Vec, nKeys uint64, sel []uint16,
 ) {
 	ht.initHash(buckets, nKeys)
 
 	for i, k := range ht.keyCols {
-		ht.rehash(buckets, i, ht.valTypes[k], keys[i], nKeys, sel)
+		ht.rehash(ctx, buckets, i, ht.valTypes[k], keys[i], nKeys, sel)
 	}
 
 	ht.finalizeHash(buckets, nKeys)
 }
 
 // buildNextChains builds the hash map from the computed hash values.
-func (ht *hashTable) buildNextChains() {
+func (ht *hashTable) buildNextChains(ctx context.Context) {
 	for id := uint64(1); id <= ht.size; id++ {
+		ht.cancelChecker.check(ctx)
 		// keyID is stored into corresponding hash bucket at the front of the next
 		// chain.
 		hash := ht.next[id]
@@ -593,7 +596,7 @@ func (builder *hashJoinBuilder) exec(ctx context.Context) {
 			builder.ht.keys[i] = builder.ht.vals[builder.ht.keyCols[i]].Slice(builder.ht.valTypes[builder.ht.keyCols[i]], batchStart, batchEnd)
 		}
 
-		builder.ht.lookupInitial(batchSize, nil)
+		builder.ht.lookupInitial(ctx, batchSize, nil)
 		nToCheck := batchSize
 
 		for nToCheck > 0 {
@@ -637,8 +640,8 @@ func (builder *hashJoinBuilder) distinctExec(ctx context.Context) {
 
 	// builder.ht.next is used to store the computed hash value of each key.
 	builder.ht.next = make([]uint64, builder.ht.size+1)
-	builder.ht.computeBuckets(builder.ht.next[1:], keyCols, builder.ht.size, nil)
-	builder.ht.buildNextChains()
+	builder.ht.computeBuckets(ctx, builder.ht.next[1:], keyCols, builder.ht.size, nil)
+	builder.ht.buildNextChains(ctx)
 }
 
 // hashJoinProber is used by the hashJoinEqOp during the probe phase. It
@@ -785,7 +788,7 @@ func (prober *hashJoinProber) exec(ctx context.Context) {
 
 			// Initialize groupID with the initial hash buckets and toCheck with all
 			// applicable indices.
-			prober.ht.lookupInitial(batchSize, sel)
+			prober.ht.lookupInitial(ctx, batchSize, sel)
 			nToCheck := batchSize
 
 			var nResults uint16
@@ -823,8 +826,8 @@ func (prober *hashJoinProber) exec(ctx context.Context) {
 // lookupInitial finds the corresponding hash table buckets for the equality
 // column of the batch and stores the results in groupID. It also initializes
 // toCheck with all indices in the range [0, batchSize).
-func (ht *hashTable) lookupInitial(batchSize uint16, sel []uint16) {
-	ht.computeBuckets(ht.buckets, ht.keys, uint64(batchSize), sel)
+func (ht *hashTable) lookupInitial(ctx context.Context, batchSize uint16, sel []uint16) {
+	ht.computeBuckets(ctx, ht.buckets, ht.keys, uint64(batchSize), sel)
 	for i := uint16(0); i < batchSize; i++ {
 		ht.groupID[i] = ht.first[ht.buckets[i]]
 		ht.toCheck[i] = i