--- a/pkg/sql/exec/sort_tmpl.go
+++ b/pkg/sql/exec/sort_tmpl.go
@@ -25,6 +25,7 @@ package exec
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 
 	"github.com/cockroachdb/apd"
@@ -87,9 +88,10 @@ func newSingleSorter(t types.T, dir distsqlpb.Ordering_Column_Direction) (colSor
 // {{range .Overloads}} {{/* for each direction */}}
 
 type sort_TYPE_DIROp struct {
-	sortCol      []_GOTYPE
-	order        []uint64
-	workingSpace []uint64
+	sortCol       []_GOTYPE
+	order         []uint64
+	workingSpace  []uint64
+	cancelChecker CancelChecker
 }
 
 func (s *sort_TYPE_DIROp) init(col coldata.Vec, order []uint64, workingSpace []uint64) {
@@ -98,9 +100,9 @@ func (s *sort_TYPE_DIROp) init(col coldata.Vec, order []uint64, workingSpace []u
 	s.workingSpace = workingSpace
 }
 
-func (s *sort_TYPE_DIROp) sort() {
+func (s *sort_TYPE_DIROp) sort(ctx context.Context) {
 	n := len(s.sortCol)
-	s.quickSort(0, n, maxDepth(n))
+	s.quickSort(ctx, 0, n, maxDepth(n))
 }
 
 func (s *sort_TYPE_DIROp) reorder() {
@@ -126,7 +128,7 @@ func (s *sort_TYPE_DIROp) reorder() {
 	}
 }
 
-func (s *sort_TYPE_DIROp) sortPartitions(partitions []uint64) {
+func (s *sort_TYPE_DIROp) sortPartitions(ctx context.Context, partitions []uint64) {
 	if len(partitions) < 1 {
 		panic(fmt.Sprintf("invalid partitions list %v", partitions))
 	}
@@ -142,7 +144,7 @@ func (s *sort_TYPE_DIROp) sortPartitions(partitions []uint64) {
 		s.order = order[partitionStart:partitionEnd]
 		s.sortCol = sortCol[partitionStart:partitionEnd]
 		n := int(partitionEnd - partitionStart)
-		s.quickSort(0, n, maxDepth(n))
+		s.quickSort(ctx, 0, n, maxDepth(n))
 	}
 }