@yuzefovich 2019-04-23
@jordanlewis you were right about probably my Mac being noisy. The results of running the benchmark on gce worker:
```
name                        old time/op    new time/op    delta
Sort/rows=2048/cols=1-24      45.6µs ± 2%    47.2µs ± 3%  +3.33%  (p=0.000 n=10+9)
Sort/rows=2048/cols=2-24       236µs ± 2%     239µs ± 2%  +1.31%  (p=0.029 n=10+10)
Sort/rows=2048/cols=4-24       412µs ± 1%     419µs ± 2%  +1.57%  (p=0.000 n=9+8)
Sort/rows=16384/cols=1-24      438µs ± 1%     469µs ± 8%  +7.16%  (p=0.013 n=9+10)
Sort/rows=16384/cols=2-24     2.00ms ± 2%    1.99ms ± 2%    ~     (p=0.515 n=8+10)
Sort/rows=16384/cols=4-24     3.38ms ± 2%    3.47ms ± 1%  +2.70%  (p=0.000 n=10+9)
Sort/rows=262144/cols=1-24    6.53ms ± 1%    6.51ms ± 1%    ~     (p=0.684 n=10+10)
Sort/rows=262144/cols=2-24    28.1ms ± 2%    28.3ms ± 3%    ~     (p=0.356 n=10+9)
Sort/rows=262144/cols=4-24    63.8ms ± 3%    62.7ms ± 3%    ~     (p=0.089 n=10+10)

name                        old speed      new speed      delta
Sort/rows=2048/cols=1-24     359MB/s ± 2%   348MB/s ± 3%  -3.23%  (p=0.000 n=10+9)
Sort/rows=2048/cols=2-24     139MB/s ± 2%   137MB/s ± 2%  -1.29%  (p=0.025 n=10+10)
Sort/rows=2048/cols=4-24     159MB/s ± 1%   155MB/s ± 6%  -2.31%  (p=0.000 n=9+9)
Sort/rows=16384/cols=1-24    299MB/s ± 1%   281MB/s ± 8%  -6.25%  (p=0.013 n=9+10)
Sort/rows=16384/cols=2-24    131MB/s ± 2%   132MB/s ± 2%    ~     (p=0.515 n=8+10)
Sort/rows=16384/cols=4-24    155MB/s ± 2%   151MB/s ± 1%  -2.64%  (p=0.000 n=10+9)
Sort/rows=262144/cols=1-24   321MB/s ± 1%   322MB/s ± 1%    ~     (p=0.684 n=10+10)
Sort/rows=262144/cols=2-24   149MB/s ± 2%   148MB/s ± 4%    ~     (p=0.368 n=10+9)
Sort/rows=262144/cols=4-24   131MB/s ± 3%   134MB/s ± 3%    ~     (p=0.089 n=10+10)

name                        old alloc/op   new alloc/op   delta
Sort/rows=2048/cols=1-24      73.2kB ± 0%    73.3kB ± 0%  +0.04%  (p=0.000 n=10+10)
Sort/rows=2048/cols=2-24       113kB ± 0%     113kB ± 0%  +0.03%  (p=0.000 n=10+10)
Sort/rows=2048/cols=4-24       206kB ± 0%     206kB ± 0%  +0.02%  (p=0.000 n=10+8)
Sort/rows=16384/cols=1-24      885kB ± 0%     885kB ± 0%  +0.00%  (p=0.000 n=9+10)
Sort/rows=16384/cols=2-24     1.52MB ± 0%    1.52MB ± 0%  +0.00%  (p=0.000 n=10+10)
Sort/rows=16384/cols=4-24     2.78MB ± 0%    2.78MB ± 0%  +0.00%  (p=0.000 n=10+10)
Sort/rows=262144/cols=1-24    16.0MB ± 0%    16.0MB ± 0%  +0.00%  (p=0.000 n=9+10)
Sort/rows=262144/cols=2-24    28.0MB ± 0%    28.0MB ± 0%  +0.00%  (p=0.000 n=10+10)
Sort/rows=262144/cols=4-24    51.5MB ± 0%    51.5MB ± 0%  +0.00%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
Sort/rows=2048/cols=1-24        27.0 ± 0%      28.0 ± 0%  +3.70%  (p=0.000 n=10+10)
Sort/rows=2048/cols=2-24        45.0 ± 0%      46.0 ± 0%  +2.22%  (p=0.000 n=10+10)
Sort/rows=2048/cols=4-24        81.0 ± 0%      82.0 ± 0%  +1.23%  (p=0.000 n=10+10)
Sort/rows=16384/cols=1-24       65.0 ± 0%      66.0 ± 0%  +1.54%  (p=0.000 n=10+10)
Sort/rows=16384/cols=2-24        121 ± 0%       122 ± 0%  +0.83%  (p=0.000 n=10+10)
Sort/rows=16384/cols=4-24        233 ± 0%       234 ± 0%  +0.43%  (p=0.000 n=10+10)
Sort/rows=262144/cols=1-24       564 ± 0%       565 ± 0%  +0.18%  (p=0.000 n=10+10)
Sort/rows=262144/cols=2-24     1.12k ± 0%     1.12k ± 0%  +0.09%  (p=0.000 n=10+10)
Sort/rows=262144/cols=4-24     2.23k ± 0%     2.23k ± 0%  +0.04%  (p=0.000 n=10+10)
```